+++ src/main/java/gyeongdan/article/controller/ArticleController.java
@@ -1,21 +1,18 @@
 package gyeongdan.article.controller;
 
 import gyeongdan.article.domain.Article;
-import gyeongdan.article.dto.ArticleUpdateRequest;
 import gyeongdan.article.service.ArticleService;
 import gyeongdan.util.CommonResponse;
-import gyeongdan.util.annotation.AdminAuthenticated;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-@RestController()
+import java.util.List;
+
+@RestController
 @RequestMapping("/api/article")
 @RequiredArgsConstructor
 public class ArticleController {
@@ -28,32 +25,9 @@ public class ArticleController {
         return ResponseEntity.ok(new CommonResponse<>(article, "게시글 조회 성공", true));
     }
 
-
     @GetMapping("")
     public ResponseEntity<?> getArticles() {
-        List<Article> article = articleService.getValidArticles();
-        return ResponseEntity.ok(new CommonResponse<>(article, "게시글 조회 성공", true));
-    }
-
-    @AdminAuthenticated
-    @GetMapping("/admin")
-    public ResponseEntity<?> getArticlesAdmin() {
-        List<Article> article = articleService.getArticles();
-        return ResponseEntity.ok(new CommonResponse<>(article, "게시글 조회 성공", true));
-    }
-
-    // TODO: 매니저 인가 필요
-    @PostMapping("/validate")
-    public ResponseEntity<?> saveArticle(@RequestParam Long id) {
-        Article savedArticle = articleService.validateArticle(id);
-        return ResponseEntity.ok(new CommonResponse<>(savedArticle, "게시글 저장 성공", true));
-    }
-
-
-    @PostMapping("/update")
-    public ResponseEntity<?> updateArticle(@RequestBody ArticleUpdateRequest articleUpdateRequest) {
-        Long savedArticleId = articleService.updateArticle(articleUpdateRequest);
-        return ResponseEntity.ok(new CommonResponse<>(savedArticleId, "게시글 수정 성공", true));
+        List<Article> articles = articleService.getValidArticles();
+        return ResponseEntity.ok(new CommonResponse<>(articles, "게시글 조회 성공", true));
     }
 }
-
